# packages/backend/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"version": "0.1.0",
		"title": "WAFRN Backend",
		"description": "API routes for wafrn social network"
	},
	"paths": {
...
		"/api/uploadMedia": {
			"post": {
				"description": "",
				"parameters": [
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"nsfw": {
										"example": "any"
									},
									"description": {
										"example": "any"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/updateMedia": {
			"get": {
				"description": "",
				"parameters": [
					{
						"name": "id",
						"in": "query",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "NSFW",
						"in": "query",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "description",
						"in": "query",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
		"/api/v2/notificationsScroll": {
			"get": {
				"description": "",
				"parameters": [
					{
						"name": "req",
						"in": "query",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
		"/api/v2/notificationsCount": {
			"get": {
				"description": "",
				"parameters": [
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
...
		"/api/mute": {
			"post": {
				"description": "",
				"parameters": [
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"userId": {
										"example": "any"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/unmute": {
			"post": {
				"description": "",
				"parameters": [
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"userId": {
										"example": "any"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/myMutes": {
			"get": {
				"description": "",
				"parameters": [
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
		"/api/unmute-user": {
			"post": {
				"description": "",
				"parameters": [
					{
						"name": "id",
						"in": "query",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
...
	},
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	}
}